Moving data shift analyzer logic into a class in its own file.
